Use ImportBeanDefintionRegistrar directly instead of import selector
Closes gh-16679
