+++ spring-boot-project/spring-boot/src/main/java/org/springframework/boot/context/properties/EnableConfigurationProperties.java
@@ -36,7 +36,8 @@ import org.springframework.context.annotation.Import;
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@Import(EnableConfigurationPropertiesImportSelector.class)
+@Import({ ConfigurationPropertiesBeanRegistrar.class,
+		ConfigurationPropertiesBindingPostProcessorRegistrar.class })
 public @interface EnableConfigurationProperties {
 
 	/**
